Initial working example for a PubSub Manager, aka Qless-native Events
This roughly follows the Python (and to a lesser extent Ruby) API, where `Client` exposes an `events` property which has `on(name, callback)` and `listen()` methods.

I noticed the existing Watchdog PubSub consumer does a clone/disconnect/reconnect, but that method does not seem to reliably give a new connection (to allow the callback to use `Client` for other calls).

The current implementation also doesn't offer the priority functionality of the existing callback based "Events" system - I thought it more important to keep parity with the upstream feature.

I had to make a guess about naming too, it's not clear to me whether this class should be in the root `Qless` namespace, or even if you have some better name than `PubSub Manager` - EventsManager is taken and I assume you wouldn't just drop the old implementation.

I've done some very minimal tests of this and it works as expected across processes. It'd be great to get some feedback in terms of naming/structure/etc and I can refine it from there.